editCompany @Min @Max

+++ src/main/kotlin/com/info/info_v1_backend/domain/company/business/dto/request/company/CompanyInformationRequest.kt
@@ -24,7 +24,7 @@ data class CompanyInformationRequest(
             this.representative,
             this.establishedAt,
             this.workerCount,
-            this.annualSales.toLong()
+            this.annualSales
         )
     }
 }
